doctors are able to input medicine for patient prescription

+++ src/main/java/org/pahappa/systems/models/Doctor.java
@@ -24,9 +24,9 @@ public class Doctor extends User {
 
     // Constructor must include staffStatus too if you use it
     public Doctor(Rolename role, String password, String nextOfKin, String address, Gender gender,
-                  LocalDate dateOfBirth,
-                  String email, String phoneNumber, String fullName, Specialty specialization,
-                  Qualification qualification, Department department, int yearsOfExperience, Status staffStatus) {
+            LocalDate dateOfBirth,
+            String email, String phoneNumber, String fullName, Specialty specialization,
+            Qualification qualification, Department department, int yearsOfExperience, Status staffStatus) {
         super(role, password, nextOfKin, address, gender, dateOfBirth, email, phoneNumber, fullName);
         this.specialization = specialization;
         this.qualification = qualification;
@@ -115,7 +115,7 @@ public class Doctor extends User {
 
     @Override
     public int hashCode() {
-        return Integer.hashCode(getId());
+        return Long.hashCode(getId());
     }
 
 }
